+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.5
 - Optimized the sound and music APIs
 - Added the color class for the default palette
+- Added the shortcut to select a color to the resource editor
 
 ## 1.1.4
 - Fixed the index check of the playm API